removed repeated code + added useSelector,useDispatch + structured utilities folder
