+++ src/components/NavBarPortfolioLink/index.jsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { ListItemPortfolio, StyledLink } from "./NavBarPortfolioLink.styles";
-import { getThemeColors } from "../../utilities";
 
 export const NavBarPortfolioLink = ({ isPortfolio }) => {
   return (
