+++ src/utilities/formatDate.jsx
@@ -1,0 +1,4 @@
+export const formatDate = (date) => {
+  const formattedDate = new Date(date).getHours() + ":" + "00";
+  return formattedDate;
+};
