+++ src/utilities/getLocalStorageItem.jsx
@@ -1,0 +1,4 @@
+export const getLocalStorageItem = (string) => {
+  const item = JSON.parse(localStorage.getItem(string));
+  return item;
+};
