+++ src/utilities/getSmallChartLabels.jsx
@@ -1,0 +1,7 @@
+export const getSmallChartLabels = () => {
+  let labels = [];
+  for (let i = 0; i < 28; i++) {
+    labels.push("");
+  }
+  return labels;
+};
