+++ src/utilities/getThemeColors.jsx
@@ -1,0 +1,4 @@
+export const getThemeColors = () => {
+  const theme = localStorage.getItem("theme");
+  return JSON.parse(theme);
+};
